evolution.cpp: fixed bug in the declaration of J.
As pointed out by (thank you!) GitHimanshuc (https://github.com/GitHimanshuc), the declaration of the index J was inconsistent with the size of the array phi_n. We note that since most of our production runs pushed the outer boundary out of causal contact with the center of the grid, this bug was unnoticed, but does not invalidate our results.